+++ src/main/java/com/banchan/repository/ReportsRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReportsRepository extends JpaRepository<Reports, Integer> {
     Integer countByReviewId(Integer reviewId);
-    Integer countByUserIdAndReviewId(Integer userId, Integer reviewId);
+    Integer countByUserIdAndReviewId(Long userId, Integer reviewId);
     Integer countByQuestionId(Integer questionId);
-    Integer countByUserIdAndQuestionId(Integer userId, Integer questionId);
+    Integer countByUserIdAndQuestionId(Long userId, Integer questionId);
 }
